Feat: Added total scores to the database and added a scoreboard
After a game over, the total score of the player will be sent to the database and when the data is fetched, the total score will also be displayed. During the game over screen, the scores of the top 10 players will be fetched from the Firebase database and will be displayed.